requirements.txt
autopep8==1.5.7
dj-database-url==0.5.0
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
asgiref==3.7.2
axios==0.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.0
idna==3.6
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
pillow==10.2.0
Pygments==2.17.2
pytz==2024.1
requests==2.31.0
rich==13.7.1
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1